Move timeifyDate into helpers folder and import it into messages
